выделил этап prelayout
