+++ TagCloud/TagCloud/PreLayout/WordScaler.cs
@@ -2,36 +2,12 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using TagCloud.Drawing;
-using TagCloud.TextProcessing;
 
-namespace TagCloud.Layout
+namespace TagCloud.PreLayout
 {
-    internal class WordLayouter : IWordLayouter
+    internal static class WordScaler
     {
-        private readonly ICloudLayouter _layouter;
-
-        public WordLayouter(ICloudLayouter layouter)
-        {
-            _layouter = layouter;
-        }
-
-        public List<Word> Layout(IDrawerOptions options, Dictionary<string, int> wordsWithFrequency)
-        {
-            var words = GetWordsWithScaledFontSize(wordsWithFrequency, options.BaseFontSize, options.FontFamily);
-
-            using var g = Graphics.FromImage(new Bitmap(1, 1));
-            foreach (var word in words)
-            {
-                var wordSize = g.MeasureString(word.Text, word.Font).ToSize();
-                word.Rectangle = _layouter.PutNextRectangle(wordSize);
-            }
-
-            _layouter.Reset();
-            return words;
-        }
-
-        private static List<Word> GetWordsWithScaledFontSize(Dictionary<string, int> wordsWithFrequency,
+        public static List<Word> GetWordsWithScaledFontSize(Dictionary<string, int> wordsWithFrequency,
             float baseFontSize,
             FontFamily fontFamily)
         {
